requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
pytz==2018.9
